+++ fishbowl_common/__init__.py
@@ -11,7 +11,15 @@ from fishbowl_common.UpdateChecker import (
     UpdateCheckResult,
 )
 from fishbowl_common.UpdateCoordinator import UpdateCoordinator, UpdateDisplay
-from fishbowl_common.UpdateDownloader import UpdateDownloader
+from fishbowl_common.UpdateDownloader import (
+    DOWNLOAD_ERROR_DIGEST,
+    DOWNLOAD_ERROR_HTTP,
+    DOWNLOAD_ERROR_IO,
+    DOWNLOAD_ERROR_NETWORK,
+    DOWNLOAD_ERROR_NO_DIGEST,
+    DOWNLOAD_ERROR_SIZE,
+    UpdateDownloader,
+)
 from fishbowl_common.UpdateInstaller import UpdateInstaller
 from fishbowl_common.version_utils import compare_versions, parse_version
 
@@ -21,6 +29,12 @@ __all__ = [
     "CHECK_ERROR_NETWORK",
     "CHECK_ERROR_RATE_LIMITED",
     "CHECK_ERROR_RESPONSE",
+    "DOWNLOAD_ERROR_DIGEST",
+    "DOWNLOAD_ERROR_HTTP",
+    "DOWNLOAD_ERROR_IO",
+    "DOWNLOAD_ERROR_NETWORK",
+    "DOWNLOAD_ERROR_NO_DIGEST",
+    "DOWNLOAD_ERROR_SIZE",
     "PatchNotes",
     "ReleaseAsset",
     "SettingsRepository",